Revert "Splash"
