+++ frontend/package.json
@@ -12,8 +12,7 @@
     "react-scripts": "^3.0.1",
     "redux": "^4.0.2",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.22.3"
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
