+++ frontend/src/components/main/main_page.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+class MainPage extends React.Component {
+
+  render() {
+    return (
+      <div>
+        <h1>DungeonChat</h1>
+        <footer>
+          Copyright &copy; 2019 Turd Burglers
+        </footer>
+      </div>
+    );
+  }
+};
+
+export default MainPage;
