--- frontend/src/components/main/main_page.jsx
@@ -1,1 +1,0 @@
-//TBD if needed
