--- frontend/src/components/main/main_page_container.js
@@ -1,1 +1,0 @@
-//TBD if needed
