+++ frontend/src/components/nav/navbar_container.js
@@ -1,6 +1,5 @@
 import { connect } from "react-redux";
 import { logout } from "../../actions/session_actions";
-import { withRouter } from 'react-router-dom';
 
 import NavBar from "./navbar";
 
@@ -8,4 +7,4 @@ const mSP = state => ({
   loggedIn: state.session.isAuthenticated
 });
 
-export default withRouter(connect(mSP, { logout })(NavBar));
+export default connect(mSP, { logout })(NavBar);
