+++ frontend/src/index.js
@@ -8,9 +8,6 @@ import jwt_decode from "jwt-decode";
 import { setAuthToken } from "./util/session_api_util";
 import { logout } from "./actions/session_actions";
 
-//Styling import
-import "./styles/output.css";
-
 document.addEventListener("DOMContentLoaded", () => {
   let store;
