--- frontend/src/styles/input.scss
@@ -1,13 +1,0 @@
-@import './reset.scss';
-@import './splash.scss';
-@import './nav.scss';
-@import './session-form.scss';
-
-//Update URL if font changes
-@import url('https://fonts.googleapis.com/css?family=Roboto&display=swap');
-
-body{
-    background-color: rgba(27, 30, 38, 1);
-    font-family: 'Roboto', sans-serif;
-    color: white;
-}
