--- frontend/src/styles/nav.scss
@@ -1,37 +1,0 @@
-.nav-bar{
-    display: flex;
-    justify-content: space-between;
-    margin: 15px;
-}
-
-.nav-right{
-    a{
-        color: white;
-        text-decoration: none;
-        span{
-            padding: 10px;
-        }
-    }
-}
-
-.nav-left{
-    a{
-        font-size: 12px;
-        color: white;
-        text-decoration: none;
-        padding: 5px 10px;
-    }
-    .btn{
-        font-size: 12px;
-        border: .5px solid rgb(15, 21, 85);
-        background:rgb(15, 21, 85);
-        color: white;
-        border-radius: 3px;
-        text-decoration: none;
-        padding: 5px 10px;
-    }   
-}
-
-.null-div{
-    display: none;
-}
