--- frontend/src/styles/output.css
@@ -1,223 +1,0 @@
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-@import url("https://fonts.googleapis.com/css?family=Roboto&display=swap");
-html, body, div, span, applet, object, iframe,
-h1, h2, h3, h4, h5, h6, p, blockquote, pre,
-a, abbr, acronym, address, big, cite, code,
-del, dfn, em, img, ins, kbd, q, s, samp,
-small, strike, strong, sub, sup, tt, var,
-b, u, i, center,
-dl, dt, dd, ol, ul, li,
-fieldset, form, label, legend,
-table, caption, tbody, tfoot, thead, tr, th, td,
-article, aside, canvas, details, embed,
-figure, figcaption, footer, header, hgroup,
-menu, nav, output, ruby, section, summary,
-time, mark, audio, video, input {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline; }
-
-/* HTML5 display-role reset for older browsers */
-article, aside, details, figcaption, figure,
-footer, header, hgroup, menu, nav, section {
-  display: block; }
-
-body {
-  line-height: 1; }
-
-ol, ul {
-  list-style: none; }
-
-blockquote, q {
-  quotes: none; }
-
-blockquote:before, blockquote:after,
-q:before, q:after {
-  content: '';
-  content: none; }
-
-table {
-  border-collapse: collapse;
-  border-spacing: 0; }
-
-.splash-wrapper {
-  z-index: 1; }
-
-.hero-background {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: -1; }
-
-.hero-top {
-  margin: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center; }
-
-.hero-left {
-  padding: 20px; }
-  .hero-left h1 {
-    font-size: 32px; }
-  .hero-left p {
-    font-size: 12px; }
-
-.hero-bottom {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 100px; }
-  .hero-bottom #signup-btn {
-    font-size: 12px;
-    border: 0.5px solid #0f1555;
-    background: #0f1555;
-    color: white;
-    border-radius: 3px;
-    text-decoration: none;
-    padding: 5px 10px;
-    width: 100px;
-    height: 25px;
-    margin: 5px;
-    text-align: center;
-    line-height: 25px; }
-  .hero-bottom #browse-btn {
-    font-size: 12px;
-    border: .5px solid white;
-    background: transparent;
-    color: white;
-    border-radius: 3px;
-    text-decoration: none;
-    padding: 5px 10px;
-    width: 100px;
-    height: 25px;
-    margin: 5px;
-    text-align: center;
-    line-height: 25px; }
-
-.splash-main {
-  position: absolute;
-  top: 500px; }
-
-.nav-bar {
-  display: flex;
-  justify-content: space-between;
-  margin: 15px; }
-
-.nav-right a {
-  color: white;
-  text-decoration: none; }
-  .nav-right a span {
-    padding: 10px; }
-
-.nav-left a {
-  font-size: 12px;
-  color: white;
-  text-decoration: none;
-  padding: 5px 10px; }
-.nav-left .btn {
-  font-size: 12px;
-  border: 0.5px solid #0f1555;
-  background: #0f1555;
-  color: white;
-  border-radius: 3px;
-  text-decoration: none;
-  padding: 5px 10px; }
-
-.null-div {
-  display: none; }
-
-#session-nav {
-  background: linear-gradient(to left, #84807d, #1b1e26);
-  display: flex;
-  height: 41px;
-  justify-content: center; }
-
-#session-nav div.container {
-  display: flex;
-  flex-direction: row;
-  width: 700px;
-  align-items: center;
-  color: white; }
-
-#session-nav div.back {
-  cursor: pointer;
-  opacity: 0.6;
-  padding-left: 10px;
-  padding-right: 10px;
-  display: flex;
-  align-items: center; }
-
-#session-nav a {
-  color: white; }
-
-#session-page {
-  display: flex;
-  justify-content: center; }
-
-.session-form-container {
-  width: 1111px;
-  display: flex;
-  flex-direction: column;
-  margin-top: 50px; }
-
-#session-page div.input {
-  width: 1111px;
-  display: inline-flex;
-  margin-top: 30px;
-  box-shadow: 0 3px 10px 1px rgba(0, 0, 0, 0.3);
-  height: 55px; }
-
-#session-page div.input-title {
-  font-size: 20px;
-  color: #707070;
-  width: 50px;
-  align-self: center;
-  padding: 0.7em 0.3em;
-  overflow: initial;
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  /* text-align: center; */ }
-
-#session-page div.input-field {
-  display: flex;
-  flex: 1;
-  background-color: #3b4045;
-  padding: 0.3em;
-  /* overflow: hidden; */
-  /* position: relative; */
-  width: 614px; }
-
-#session-page div.input-field input {
-  width: 100%;
-  padding: 0 10px; }
-
-.title-container {
-  background-color: #2D2F30;
-  display: flex;
-  justify-content: center; }
-
-div.bar {
-  height: 28px;
-  width: 0px;
-  border-left: none;
-  border-top: none;
-  border-bottom: none;
-  border-style: solid;
-  border-width: 1px;
-  border-color: grey;
-  transform: translate(1px, 13px); }
-
-body {
-  background-color: #1b1e26;
-  font-family: 'Roboto', sans-serif;
-  color: white; }
-
-/*# sourceMappingURL=output.css.map */
