+++ frontend/src/util/route_util.js
@@ -7,7 +7,7 @@ const Auth = ({ component: Component, path, loggedIn, exact }) => (
     path={path}
     exact={exact}
     render={props => (
-      !loggedIn ? <Component {...props} /> : <Redirect to="/profile" />
+      !loggedIn ? <Component {...props} /> : <Redirect to="/tweets" />
     )}
   />
 );
